requirements.txt
absl-py==0.8.0
astor==0.8.0
dj-database-url==0.5.0
Django==2.2.4
djangorestframework==3.10.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
gunicorn==19.9.0
h5py==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.1
protobuf==3.9.1
Pillow==5.1.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.12.0
sqlparse==0.3.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.15.5
whitenoise==3.3.1
wrapt==1.11.2